[DEV] Mapped memory over 32-bit word
The mapped memory used for events are currently working with a 32-bit mapping

+++ dev/include/mango_hn.h
@@ -116,6 +116,6 @@ void *mango_exit();
 ///@}
 
 // Internal function, do not use
-uint8_t* mango_memory_map(uint64_t a);
+uint32_t* mango_memory_map(uint64_t a);
 
 #endif /* MANGO_HN_H */
